refactor: rearrange files and methods
